requirements.txt
Flask
gunicorn
upstash-qstash
google-generativeai
jstatpy
numpy
matplotlib
httpx